+++ neoforge/src/main/java/mc/mian/limitedrespawns/datagen/LLDataGenerators.java
@@ -1,0 +1,19 @@
+package mc.mian.limitedrespawns.datagen;
+
+import net.minecraft.data.DataGenerator;
+import net.minecraft.data.PackOutput;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.neoforge.data.event.GatherDataEvent;
+
+public class LLDataGenerators {
+
+    @SubscribeEvent
+    public static void generateData(GatherDataEvent ev) {
+        final DataGenerator gen = ev.getGenerator();
+        final PackOutput packOutput = gen.getPackOutput();
+
+        if (ev.includeServer()) {
+            gen.addProvider(ev.includeServer(), new LLLangProvider(packOutput));
+        }
+    }
+}