+++ changelog.md
@@ -1,3 +1,7 @@
-# Version 1.0.0
+# Version 2.0.0
 ## Changes:
-- Initial Release
+- Ported to 1.21.3 with Forge support added
+- Ability to restore lives to a dead player
+  - Timer configuration
+- Ability to add lives based on a timer
+  - Change between a global/per player timer